fix(desktop): address v2 DiffPane review feedback (a11y, error text, expand-unchanged scope)
- Move "Show all lines" out of the per-file metadata into the pane header
  (DiffPaneHeaderExtras) and persist via useSettings.expandUnchanged.
  Pierre's CodeView treats expandUnchanged as pane-wide, so a per-file
  Eye button in the file header was misleading — toggling on file B
  silently affected file A. The pane-header button now matches the API
  semantics, and the setting persists across sessions like diffStyle /
  showDiffComments.
- Reveal the per-file Discard button on :focus-within too, not just
  :hover, so keyboard users can see the destructive control before
  pressing it.
- Add select-text / cursor-text to the diff "Loading…" / "Unable to load
  diff" container, per the apps/desktop convention that user-visible
  error text must be copyable.
